Extract HTTPClient tests into more classes
To improve readability
