+++ src/xcode/ENA/ENA/Source/Client/HTTP Client/__tests__/HTTPClient+RegistrationTokenTests.swift
@@ -1,0 +1,182 @@
+//
+// Corona-Warn-App
+//
+// SAP SE and all other contributors
+// copyright owners license this file to you under the Apache
+// License, Version 2.0 (the "License"); you may not use this
+// file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+//
+
+@testable import ENA
+import Foundation
+import XCTest
+
+final class HTTPClientRegistrationTokenTests: XCTestCase {
+	private let expectationsTimeout: TimeInterval = 2
+
+	func testGetRegistrationToken_TeleTANSuccess() throws {
+		let expectedToken = "SomeToken"
+		let stack = MockNetworkStack(
+			httpStatus: 200,
+			responseData: try? JSONEncoder().encode(GetRegistrationTokenResponse(registrationToken: expectedToken))
+		)
+
+		let successExpectation = expectation(
+			description: "Expect that we got a completion"
+		)
+
+		HTTPClient.makeWith(mock: stack).getRegistrationToken(forKey: "1234567890", withType: "TELETAN") { result in
+			defer { successExpectation.fulfill() }
+			switch result {
+			case .success(let token):
+				XCTAssertEqual(token, expectedToken)
+			case .failure:
+				XCTFail("Encountered Error when receiving registration token!")
+			}
+		}
+		waitForExpectations(timeout: expectationsTimeout)
+	}
+
+	func testGetRegistrationToken_GUIDSuccess() throws {
+		let expectedToken = "SomeToken"
+		let stack = MockNetworkStack(
+			httpStatus: 200,
+			responseData: try? JSONEncoder().encode(GetRegistrationTokenResponse(registrationToken: expectedToken))
+		)
+
+		let successExpectation = expectation(
+			description: "Expect that we got a completion"
+		)
+
+		HTTPClient.makeWith(mock: stack).getRegistrationToken(forKey: "1234567890", withType: "GUID") { result in
+			defer { successExpectation.fulfill() }
+			switch result {
+			case .success(let token):
+				XCTAssertEqual(token, expectedToken)
+			case .failure:
+				XCTFail("Encountered Error when receiving registration token!")
+			}
+		}
+		waitForExpectations(timeout: expectationsTimeout)
+	}
+
+	func testGetRegistrationToken_TANAlreadyUsed() throws {
+		let stack = MockNetworkStack(
+			httpStatus: 400,
+			responseData: Data()
+		)
+
+		let successExpectation = expectation(
+			description: "Expect that we got a completion"
+		)
+
+		HTTPClient.makeWith(mock: stack).getRegistrationToken(forKey: "1234567890", withType: "TELETAN") { result in
+			defer { successExpectation.fulfill() }
+			switch result {
+			case .success:
+				XCTFail("Backend returned 400 - the request should have failed!")
+			case .failure(let error):
+				switch error {
+				case .teleTanAlreadyUsed:
+					break
+				default:
+					XCTFail("The error was not .teleTanAlreadyUsed!")
+				}
+			}
+		}
+		waitForExpectations(timeout: expectationsTimeout)
+	}
+
+	func testGetRegistrationToken_GUIDAlreadyUsed() throws {
+		let stack = MockNetworkStack(
+			httpStatus: 400,
+			responseData: Data()
+		)
+
+		let successExpectation = expectation(
+			description: "Expect that we got a completion"
+		)
+
+		HTTPClient.makeWith(mock: stack).getRegistrationToken(forKey: "1234567890", withType: "GUID") { result in
+			defer { successExpectation.fulfill() }
+			switch result {
+			case .success:
+				XCTFail("Backend returned 400 - the request should have failed!")
+			case .failure(let error):
+				switch error {
+				case .qRAlreadyUsed:
+					break
+				default:
+					XCTFail("The error was not .qRAlreadyUsed!")
+				}
+			}
+		}
+		waitForExpectations(timeout: expectationsTimeout)
+	}
+
+	func testGetRegistrationToken_MalformedResponse() throws {
+		let stack = MockNetworkStack(
+			httpStatus: 200,
+			responseData: Data(bytes: [0xA, 0xB], count: 2)
+		)
+
+		let successExpectation = expectation(
+			description: "Expect that we got a completion"
+		)
+
+		HTTPClient.makeWith(mock: stack).getRegistrationToken(forKey: "1234567890", withType: "GUID") { result in
+			defer { successExpectation.fulfill() }
+			switch result {
+			case .success:
+				XCTFail("Backend returned random bytes - the request should have failed!")
+			case .failure(let error):
+				switch error {
+				case .invalidResponse:
+					break
+				default:
+					XCTFail("The error was not .invalidResponse!")
+				}
+			}
+		}
+		waitForExpectations(timeout: expectationsTimeout)
+	}
+
+	func testGetRegistrationToken_MalformedJSONResponse() throws {
+		let stack = MockNetworkStack(
+			httpStatus: 200,
+			responseData: """
+			{ "NotregistrationToken":"Hello" }
+			""".data(using: .utf8)
+		)
+
+		let successExpectation = expectation(
+			description: "Expect that we got a completion"
+		)
+
+		HTTPClient.makeWith(mock: stack).getRegistrationToken(forKey: "1234567890", withType: "GUID") { result in
+			defer { successExpectation.fulfill() }
+			switch result {
+			case .success:
+				XCTFail("Backend returned 400 - the request should have failed!")
+			case .failure(let error):
+				switch error {
+				case .invalidResponse:
+					break
+				default:
+					XCTFail("The error was not .invalidResponse!")
+				}
+			}
+		}
+		waitForExpectations(timeout: expectationsTimeout)
+	}
+}
